feat(app): 7tv paints
