+++ lefthook.yml
@@ -1,7 +1,7 @@
 commit-msg:
   commands:
     'lint commit message':
-      run: pnpm commitlint --edit {1}
+      run: bun commitlint --edit {1}
 
 pre-commit:
   parallel: true
