+++ src/components/Chat/components/ChatMessage/CosmeticUsername.stories.tsx
@@ -1,4 +1,4 @@
-import type { PaintData } from '@app/services/ws/seventv-ws-service';
+import type { PaintData } from '@app/utils/color/seventv-ws-service';
 import type { Meta, StoryObj } from '@storybook/react';
 import { View } from 'react-native';
 import { PaintedUsername } from './CosmeticUsername';
